requirements.txt
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0  # optional but recommended for environment variables